#41007 Split AdminSearchService back out as separate class - to keep the admin search logic separate
